fix: scope mobile drift checks by upstream batch
