+++ docs/superpowers/plans/2026-05-04-upstream-rebase-process-01-manifest-foundation.md
@@ -224,7 +224,7 @@ upstream-postrebase-audit:
 
 .PHONY: mobile-drift-rebase-check
 mobile-drift-rebase-check:
-	$(UPSTREAM_PREFLIGHT) run mobile-drift-check
+	$(UPSTREAM_PREFLIGHT) run mobile-drift-check $(if $(BATCH),-- --batch $(BATCH),)
 
 .PHONY: ci-invariants-check
 ci-invariants-check:
